requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.8.1
asttokens==2.4.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==6.1.0
blinker==1.8.2
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
check-wheel-contents==0.6.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
Cython==3.0.11
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
Django==5.1.1
djangorestframework==3.15.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.21.2
executing==2.1.0
eyeD3==0.8.10
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
grpcio==1.30.2
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.11.3
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jsonify==0.5
jsonschema==3.2.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nemo-emblems==5.6.0
netaddr==0.8.0
netifaces==0.11.0
nh3==0.2.18
oauthlib==3.2.0
onboard==1.4.1
packaging==21.3
PAM==0.4.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2024.0.2
pipreqs==0.5.0
pkginfo==1.10.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pydantic==2.9.2
pydantic_core==2.23.4
pyelftools==0.27
Pygments==2.18.0
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.9.0.post0
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pytz==2024.2
pyxdg==0.27
PyYAML==6.0.2
pyzmq==26.2.0
readme_renderer==44.0
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.8.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
sqlparse==0.5.1
stack-data==0.6.3
systemd-python==234
texttable==1.6.4
tinycss2==1.1.1
tldextract==3.1.2
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
twine==5.1.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
virtualenv==20.26.5
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==3.0.4
wheel-filename==1.4.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
yarg==0.1.9
youtube-dl==2021.12.17
zipp==3.8.0
